refactor: 사용자정의 FamilyRepository 삭제 #29
