+++ src/main/java/server/acode/domain/family/repository/FragranceFamilyRepositoryCustom.java
@@ -4,6 +4,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 import server.acode.domain.family.dto.SimilarFragranceOrCond;
+import server.acode.domain.family.dto.request.FragranceFilterCond;
+import server.acode.domain.family.dto.response.DisplayFragrance;
+import server.acode.domain.family.dto.response.HomeFragrance;
 import server.acode.domain.fragrance.dto.request.SearchCond;
 import server.acode.domain.fragrance.dto.response.ExtractFragrance;
 import server.acode.domain.fragrance.dto.response.FamilyCountDto;
@@ -13,6 +16,10 @@ import java.util.List;
 
 @Repository
 public interface FragranceFamilyRepositoryCustom {
+    List<HomeFragrance> search(String familyName);
+
+    Page<DisplayFragrance> searchByFilter(FragranceFilterCond cond, String additionalFamily, Pageable pageable);
+
     List<Long> searchFamilyByFragranceId(Long fragranceId);
 
     List<FragranceInfo> searchSimilarFragranceAnd(Long fragranceId, Long familyId1, Long familyId2);
