+++ src/main/java/server/acode/domain/family/repository/FragranceFamilyRepositoryImpl.java
@@ -1,15 +1,22 @@
 package server.acode.domain.family.repository;
 
+import com.querydsl.core.QueryResults;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
 import org.springframework.stereotype.Repository;
 import server.acode.domain.family.dto.SimilarFragranceOrCond;
+import server.acode.domain.family.dto.request.FragranceFilterCond;
+import server.acode.domain.family.dto.response.DisplayFragrance;
+import server.acode.domain.family.dto.response.HomeFragrance;
+import server.acode.domain.family.dto.response.QDisplayFragrance;
+import server.acode.domain.family.dto.response.QHomeFragrance;
 import server.acode.domain.family.entity.QFragranceFamily;
 import server.acode.domain.fragrance.dto.request.SearchCond;
 import server.acode.domain.fragrance.dto.response.*;
@@ -20,6 +27,7 @@ import java.util.List;
 import static org.springframework.util.StringUtils.*;
 import static server.acode.domain.family.entity.QFamily.*;
 import static server.acode.domain.family.entity.QFragranceFamily.*;
+import static server.acode.domain.fragrance.entity.QBrand.brand;
 import static server.acode.domain.fragrance.entity.QFragrance.fragrance;
 
 @Repository
@@ -28,6 +36,60 @@ public class FragranceFamilyRepositoryImpl implements FragranceFamilyRepositoryC
 
     public FragranceFamilyRepositoryImpl(EntityManager em) {
         this.queryFactory = new JPAQueryFactory(em);
+    }
+
+    @Override
+    public List<HomeFragrance> search(String familyName) {
+        return queryFactory
+                .select(new QHomeFragrance(
+                        fragrance.id.as("fragranceId"),
+                        fragrance.name.as("fragranceName"),
+                        brand.korName.as("korBrand"),
+                        fragrance.style,
+                        fragrance.poster
+                ))
+                .from(fragranceFamily)
+                .join(fragranceFamily.family, family)
+                .join(fragranceFamily.fragrance, fragrance)
+                .join(fragrance.brand, brand)
+                .where(family.korName.eq(familyName),
+                        fragrance.poster.isNotNull().and(fragrance.poster.ne(""))
+                )
+                .orderBy(fragrance.view.desc())
+                .limit(6)
+                .fetch();
+
+    }
+
+    @Override
+    public Page<DisplayFragrance> searchByFilter(FragranceFilterCond cond, String additionalFamily, Pageable pageable) {
+
+        //TODO 카운트 쿼리 분리
+        QueryResults<DisplayFragrance> results = queryFactory
+                .select(new QDisplayFragrance(
+                        fragrance.id.as("fragranceId"),
+                        brand.korName.as("brandName"),
+                        fragrance.name.as("fragranceName"),
+                        fragrance.thumbnail
+                ))
+                .from(fragranceFamily)
+                .join(fragranceFamily.fragrance, fragrance)
+                .join(fragranceFamily.family, family)
+                .join(fragrance.brand, brand)
+                .where(brandNameEq(cond.getBrand()),
+                        familyNameEq(cond.getFamily()),
+                        additionalFamilyNameEq(additionalFamily)
+                )
+                .groupBy(fragrance.id)
+                .offset(pageable.getOffset())
+                .limit(pageable.getPageSize())
+                .fetchResults();
+
+
+        List<DisplayFragrance> content = results.getResults();
+        long total = results.getTotal();
+
+        return new PageImpl<>(content, pageable, total);
     }
 
 
@@ -216,13 +278,16 @@ public class FragranceFamilyRepositoryImpl implements FragranceFamilyRepositoryC
                 : null;
     }
 
+    private BooleanExpression brandNameEq(String brandName) {
+        return hasText(brandName)? brand.korName.eq(brandName): null;
+    }
+
     private BooleanExpression familyNameEq(String familyName) {
-        return hasText(familyName)
-                ? family.korName.eq(familyName)
-                : null;
+        return hasText(familyName) ? family.korName.eq(familyName) : null;
     }
 
     private BooleanExpression additionalFamilyNameEq(String additionalFamily) {
+        // 서브 쿼리를 위한 QEntity 추가 생성
         QFragranceFamily fragranceFamilySub = new QFragranceFamily("fragranceFamilySub");
 
         return hasText(additionalFamily)
