+++ src/main/java/server/acode/domain/family/service/HomeService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import server.acode.domain.family.dto.response.HomeFragrance;
 import server.acode.domain.family.repository.FamilyRepository;
+import server.acode.domain.family.repository.FragranceFamilyRepository;
 import server.acode.domain.ingredient.dto.response.IngredientOfTheDay;
 import server.acode.domain.ingredient.repository.IngredientRepository;
 import server.acode.global.common.ErrorCode;
@@ -19,13 +20,14 @@ import java.util.stream.Collectors;
 public class HomeService {
 
     private final FamilyRepository familyRepository;
+    private final FragranceFamilyRepository fragranceFamilyRepository;
     private final IngredientRepository ingredientRepository;
 
     public List<HomeFragrance> search(String familyName){
 
         if(!familyRepository.existsByKorName(familyName)) throw new CustomException(ErrorCode.FAMILY_NOT_FOUND);
 
-        List<HomeFragrance> result = familyRepository.search(familyName); // 계열과 일치하는 향수 가져오기
+        List<HomeFragrance> result = fragranceFamilyRepository.search(familyName); // 계열과 일치하는 향수 가져오기
         return result;
     }
