+++ src/main/java/server/acode/domain/fragrance/repository/FragranceRepositoryImpl.java
@@ -1,15 +1,26 @@
 package server.acode.domain.fragrance.repository;
 
+import com.querydsl.core.QueryResults;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
+import server.acode.domain.family.dto.response.DisplayFragrance;
+import server.acode.domain.family.dto.response.QDisplayFragrance;
 import server.acode.domain.fragrance.entity.Concentration;
 
 import java.util.List;
 
 import static org.springframework.util.StringUtils.*;
+import static server.acode.domain.fragrance.entity.QBrand.brand;
 import static server.acode.domain.fragrance.entity.QFragrance.*;
+import static server.acode.domain.ingredient.entity.QBaseNote.baseNote;
+import static server.acode.domain.ingredient.entity.QIngredient.ingredient;
+import static server.acode.domain.ingredient.entity.QMiddleNote.middleNote;
+import static server.acode.domain.ingredient.entity.QTopNote.topNote;
 
 @Repository
 public class FragranceRepositoryImpl implements FragranceRepositoryCustom {
@@ -17,6 +28,36 @@ public class FragranceRepositoryImpl implements FragranceRepositoryCustom {
 
     public FragranceRepositoryImpl(EntityManager em) {
         this.queryFactory = new JPAQueryFactory(em);
+    }
+
+
+    @Override
+    public Page<DisplayFragrance> searchByIngredient(String ingredientName, Pageable pageable) {
+
+        QueryResults<DisplayFragrance> results = queryFactory
+                .select(new QDisplayFragrance(
+                        fragrance.id.as("fragranceId"),
+                        brand.korName.as("brandName"),
+                        fragrance.name.as("fragranceName"),
+                        fragrance.thumbnail
+                ))
+                .from(fragrance)
+                .join(fragrance.brand, brand)
+                .join(topNote).on(fragrance.id.eq(topNote.fragrance.id))
+                .join(middleNote).on(fragrance.id.eq(middleNote.fragrance.id))
+                .join(baseNote).on(fragrance.id.eq(baseNote.fragrance.id))
+                .join(ingredient).on(topNote.ingredient.id.eq(ingredient.id)
+                        .or(middleNote.ingredient.id.eq(ingredient.id))
+                        .or(baseNote.ingredient.id.eq(ingredient.id))
+                )
+                .where(ingredient.korName.eq(ingredientName))
+                .groupBy(fragrance.id)
+                .fetchResults();
+
+        List<DisplayFragrance> content = results.getResults();
+        long total = results.getTotal();
+
+        return new PageImpl<>(content, pageable, total);
     }
