+++ src/main/resources/application.yml
@@ -39,8 +39,8 @@ spring:
         dialect: org.hibernate.dialect.MySQLDialect
         database-platform: org.hibernate.dialect.MySQLDialect
         default_batch_fetch_size: 100
-        show_sql: true
-        format_sql: true
+#        show_sql: true
+#        format_sql: true
 
 token:
   secret: ${SECRET_KEY}
